+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntitySearchField bug fixed, now clears selection when an empty field loses focus, only if the selection isn't empty, selection uses Notify.SET, did not play well with ConditionModel.autoEnable.
 - EntityPanel.Builder.icon() now ControlIcon based.
+- EntityComboBoxPanel.Builder.comboBoxPreferredWidth() renamed preferredComboBoxWidth().
 
 ## 0.18.49
 ### is.codion.common.core